+++ ErsatzTV.Application/MediaCollections/Commands/UpdateTraktList.cs
@@ -1,0 +1,5 @@
+using ErsatzTV.Core;
+
+namespace ErsatzTV.Application.MediaCollections;
+
+public record UpdateTraktList(int Id, bool AutoRefresh) : IRequest<Option<BaseError>>;