+++ ErsatzTV.Application/MediaCollections/Queries/GetTraktListById.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.MediaCollections;
+
+public record GetTraktListById(int Id) : IRequest<Option<TraktListViewModel>>;