+++ ErsatzTV.Core/Domain/Collection/TraktList.cs
@@ -9,5 +9,8 @@ public class TraktList
     public string Name { get; set; }
     public string Description { get; set; }
     public int ItemCount { get; set; }
+    public bool AutoRefresh { get; set; }
+    public DateTime? LastUpdate { get; set; }
+    public DateTime? LastMatch { get; set; }
     public List<TraktListItem> Items { get; set; }
 }